+++ packages/core/src/constants/category-set.js
@@ -54,6 +54,15 @@ export const CATEGORY_ID = {
   [categoryPath.photography]: '574d028748fa171000c45d48',
 }
 
+// flip CATEGORY_ID to a {id: path} object
+let categoryIDToPath = {}
+Object.entries(CATEGORY_ID).forEach(([key, value]) => {
+  categoryIDToPath[value] = key
+})
+export const GET_CATEGORY_PATH_FROM_ID = id => {
+  return categoryIDToPath[id]
+}
+
 const subcategoryPath = {
   all: 'all',
   hk: 'hk',
@@ -233,6 +242,17 @@ export const SUBCATEGORY_ID = {
   [subcategoryPath.letter]: '63206383207bf7c5f871626b',
 }
 
+// flip SUBCATEGORY_ID to a {id: path} object
+let subcategoryIDToPath = {}
+Object.entries(SUBCATEGORY_ID)
+  .filter(([key, value]) => key !== subcategoryPath.all)
+  .forEach(([key, value]) => {
+    subcategoryIDToPath[value] = key
+  })
+export const GET_SUBCATEGORY_PATH_FROM_ID = id => {
+  return subcategoryIDToPath[id]
+}
+
 export const CATEGORY_SET = {
   [categoryPath.world]: [
     subcategoryPath.all,
@@ -326,6 +346,8 @@ export default {
   CATEGORY_ORDER,
   CATEGORY_LABEL,
   CATEGORY_ID,
+  GET_CATEGORY_PATH_FROM_ID,
+  GET_SUBCATEGORY_PATH_FROM_ID,
   SUBCATEGORY_LABEL,
   SUBCATEGORY_ID,
   CATEGORY_SET,