+++ packages/universal-header/src/utils/links.js
@@ -11,10 +11,12 @@ import releaseBranchConsts from '@twreporter/core/lib/constants/release-branch'
 import forEach from 'lodash/forEach'
 import reduce from 'lodash/reduce'
 import split from 'lodash/split'
+import indexOf from 'lodash/indexOf'
 const _ = {
   forEach,
   reduce,
   split,
+  indexOf,
 }
 
 const originsForClient = origins.forClientSideRendering
@@ -70,8 +72,16 @@ export const checkReferrer = (
   referrer = '',
   releaseBranch = defaultReleaseBranch
 ) => {
-  const url = new URL(referrer)
-  return url.origin === mainBaseURL[releaseBranch]
+  try {
+    const url = new URL(referrer)
+    return url.origin === mainBaseURL[releaseBranch]
+  } catch (err) {
+    return false
+  }
+}
+
+export const checkPathnameParent = (pathname = '', parent = '') => {
+  return _.indexOf(_.split(pathname, '/'), parent) === 1
 }
 
 export const getCategoryLink = (